extract date helpers + comment FormData blob cast
Followup polish on PR #34:

1. Move formatPickupDate/formatTimeRangeAny/fmt12/parseHMAny out of
   the local function bodies in donation-details/[id].tsx and
   pickup-details/[id].tsx and into src/utils/dateHelpers.ts where
   they belong. Both files now import from there. Kills the duplication
   smell I introduced when I copied the helpers from pickup-details
   into donation-details.

2. Add a comment to the {uri, name, type} -> Blob double-cast in
   submitCompletionDetails explaining why it's there. RN's FormData
   accepts file objects in this shape but TypeScript's DOM types only
   know about Blob | string, so the cast is the standard workaround.
   No reviewer should have to wonder.

No behavior change, no new dependencies, no test impact.
